+++ src/app/notice/notice_list/group.notice.list.component.ts
@@ -1,15 +1,12 @@
-import {Component, EventEmitter} from '@angular/core';
+import {Component} from '@angular/core';
 
-import {Router, ActivatedRoute} from '@angular/router';
+import {ActivatedRoute, Router} from '@angular/router';
 import {NoticeService} from '../services/notice.service';
-import {Notice} from '../models/notice';
-
-import {FriendlyDatePipe} from '../../common/pipes/friendly.date.pipe';
-import {EllipsisPipe} from '../../common/pipes/ellipsis.pipe';
 import {AuthService} from '../../../providers/auth-service';
 import {NoticeListComponent} from './notice.list.component';
-import {MaterializeAction} from 'angular2-materialize';
-import {GroupService} from "../../group/group.service";
+import {GroupService} from '../../group/group.service';
+import {MatDialog} from '@angular/material';
+import {DialogAreYouSureComponent} from '../../common/modals/are.you.sure.component';
 
 @Component({
   selector: 'app-group-notice-list-component',
@@ -20,8 +17,8 @@ import {GroupService} from "../../group/group.service";
 export class GroupNoticeListComponent extends NoticeListComponent {
 
   notices: any[] = [];
-  groupId: string = '';
-  schoolId: string = '';
+  groupId = '';
+  schoolId = '';
   loading = true;
   groupAdmin = false;
 
@@ -29,6 +26,7 @@ export class GroupNoticeListComponent extends NoticeListComponent {
               public groupService: GroupService,
               public noticeService: NoticeService,
               public router: Router,
+              public dialog: MatDialog,
               public route: ActivatedRoute) {
 
     super(router, groupService);
@@ -121,19 +119,33 @@ export class GroupNoticeListComponent extends NoticeListComponent {
 
   deleteNotice(notice, i): void {
 
-    this.auth.getFirebaseTokenAsPromise().then(() => {
+    const dialogRef = this.dialog.open(DialogAreYouSureComponent, {
+      data: {
+        title: 'Delete this notice?'
+      }
+    });
 
-      this.noticeService.deleteNotice(notice).subscribe(res => {
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.loading = true;
 
-        this.notices.splice(i, 1);
+        this.auth.getFirebaseTokenAsPromise().then(() => {
 
-      }, error => {
+          this.noticeService.deleteNotice(notice).subscribe(res => {
 
-        console.log(error);
+            this.notices.splice(i, 1);
+            this.loading = false;
 
-      });
+          }, error => {
 
+            console.log(error);
+            this.loading = false;
 
+          });
+
+        });
+
+      }
     });
